Fix - apache unable to reliably get server's name
I don't know if this is because Fedora typically uses a FQDN for the host name
as opposed to having hostname short and using /etc/hosts to bind the host to a
domain.  Regardless, this fix is simple in that it uses a stock Fedora 12 conf
as a template and plugs in the FQDN for ServerName.

Resolves: Starting httpd: httpd: Could not reliably determine the server's
fully qualified domain name, using hector.doubledog.org for ServerName