New - selinux::boolean definition
This is just a wrapper around puppet's own selboolean resource type.  It
reads a little nicer, except when a reference is needed to a definition
instance, which still requires the use of selboolean (instead of
Selinux::boolean).  However, this may prove useful still if booleans are
to be ignored if/when SELinux is disabled.