Change - selinux::boolean value must be boolean
For some stupid reason Puppet's seboolean type requires that the value
be "on" or "off" even though the setsebool tool takes a variety of
acceptable values.  Since we wrap Puppet's dumb type, we might as well
make the wrapper smart and expect a true boolean value.

This also drops the conditional on ::selinux_simple which seems to be
returning the wrong value on brian-f20 (in enforcing mode).  Hopefully
the provider bugs have been worked out by now.