Refactor - eliminate duplicate templates
At one time this approach of having one template per OS release was very
nice to best illustrate exactly what was changing via Puppet's show_diff
reporting.  However, the file seems to be very static lately and if this
is ever needed again, a better approach would be to have a baseline
standard and exceptions, but only as needed.

Signed-off-by: John Florian <jflorian@doubledog.org>